CGlobeView 2D now pixelexact on ATI and Intel

+++ StoreManager/CGLFont.h
@@ -7,13 +7,13 @@ public:
 
 	BOOL Create(CString face, UINT size, BOOL bold, BOOL italic);
 	BOOL Create(CFont* font);
-	UINT Render(wchar_t* pStr, float x, float y, int cCount=-1);
+	UINT Render(wchar_t* pStr, int x, int y, int cCount=-1);
 	UINT GetTextWidth(wchar_t* pStr, int cCount=-1);
 	UINT GetTextHeight(wchar_t* pStr);
 
 protected:
 	BOOL Initialize(HFONT hFont);
-	float RenderChar(UCHAR ch, float x, float y, float* pHeight);
+	UINT RenderChar(UCHAR ch, int x, int y, UINT* pHeight);
 
 private:
 	float TexCoords[256-32][4];
